+++ src/TqkLibrary.Proxy/Helpers/Socks5_Request.cs
@@ -79,7 +79,7 @@ namespace TqkLibrary.Proxy.Helpers
             {
                 yield return b;
             }
-            foreach (var b in BitConverter.GetBytes(IPAddress.HostToNetworkOrder(this.DSTPORT)))
+            foreach (var b in BitConverter.GetBytes(IPAddress.HostToNetworkOrder((short)this.DSTPORT)))
             {
                 yield return b;
             }
